+++ packages/template-cli/src/domain/model/cli_response.ts
@@ -1,5 +1,5 @@
 interface ExitMessage {
-    code: number,
+    code: number | string,
     message: string
 }