+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0](https://github.com/amido/stacks-webapp-template/compare/@amido-stacks/eslint-config@0.0.6...@amido-stacks/eslint-config@0.8.0) (2020-03-04)
+
+
+### Bug Fixes
+
+* linting changes ([e65eea3](https://github.com/amido/stacks-webapp-template/commit/e65eea35f64c7530e1697d1c14e52cc748a1b5ce))
+* wrong main file ([#59](https://github.com/amido/stacks-webapp-template/issues/59)) ([425b410](https://github.com/amido/stacks-webapp-template/commit/425b410c9cfa5308460d2d8b920dd8d588a34e94))
+
+
+### Features
+
+* added custom event ([#54](https://github.com/amido/stacks-webapp-template/issues/54)) ([25e86f8](https://github.com/amido/stacks-webapp-template/commit/25e86f8187df4807b10b4d3fd122132ce0c51213))
+* unit tests in Azure Pipelines - 979 ([#25](https://github.com/amido/stacks-webapp-template/issues/25)) ([099c845](https://github.com/amido/stacks-webapp-template/commit/099c845beb8ada51888f66449521a9a768c7e07c))
+
+
+
+
+
+# [0.7.0](https://github.com/amido/stacks-webapp-template/compare/@amido-stacks/eslint-config@0.0.6...@amido-stacks/eslint-config@0.7.0) (2020-02-27)
+
+
+### Bug Fixes
+
+* wrong main file ([#59](https://github.com/amido/stacks-webapp-template/issues/59)) ([425b410](https://github.com/amido/stacks-webapp-template/commit/425b410c9cfa5308460d2d8b920dd8d588a34e94))
+
+
+### Features
+
+* added custom event ([#54](https://github.com/amido/stacks-webapp-template/issues/54)) ([25e86f8](https://github.com/amido/stacks-webapp-template/commit/25e86f8187df4807b10b4d3fd122132ce0c51213))
+* unit tests in Azure Pipelines - 979 ([#25](https://github.com/amido/stacks-webapp-template/issues/25)) ([099c845](https://github.com/amido/stacks-webapp-template/commit/099c845beb8ada51888f66449521a9a768c7e07c))
+
+
+
+
+
 # [0.6.0](https://github.com/amido/stacks-webapp-template/compare/@amido-stacks/eslint-config@0.0.6...@amido-stacks/eslint-config@0.6.0) (2020-02-27)